+++ validator-e2e/jest.config.js
@@ -1,10 +1,10 @@
 module.exports = {
-    roots: ['<rootDir>'],
-    testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$',
-    moduleFileExtensions: ['ts', 'tsx', 'js'],
-    collectCoverage: true,
-    collectCoverageFrom: ['**/*.ts'],
-    transformIgnorePatterns: [
-        'node_modules/(?!(alcaeus|hydra-validator-core)/)',
-    ],
+  roots: ['<rootDir>'],
+  testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$',
+  moduleFileExtensions: ['ts', 'tsx', 'js'],
+  collectCoverage: true,
+  collectCoverageFrom: ['**/*.ts'],
+  transformIgnorePatterns: [
+    'node_modules/(?!(alcaeus|hydra-validator-core)/)',
+  ],
 }