+++ validator-e2e/lib/steps/constraints/Constraint.ts
@@ -1,5 +1,5 @@
-import { HydraResource } from 'alcaeus/types/Resources'
-import { IHydraResponse } from 'alcaeus/types/HydraResponse'
+import { HydraResource, ResourceIndexer } from 'alcaeus'
+import { HydraResponse } from 'alcaeus/HydraResponse'
 
 export type ConstraintOperator = 'eq' | 'gt' | 'ge' | 'lt' | 'le' | 'regex' | 'function'
 export type ConstraintType = 'Representation' | 'Response' | null
@@ -31,13 +31,13 @@ export abstract class Constraint<T = unknown> {
   protected abstract sanityCheckValue(value: unknown): boolean
 }
 
-export abstract class ResponseConstraint extends Constraint<IHydraResponse> {
+export abstract class ResponseConstraint extends Constraint<HydraResponse> {
   public get type(): ConstraintType {
     return 'Response'
   }
 }
 
-export abstract class RepresentationConstraint extends Constraint<HydraResource> {
+export abstract class RepresentationConstraint extends Constraint<HydraResource & ResourceIndexer> {
   public get type(): ConstraintType {
     return 'Representation'
   }