+++ validator-e2e/lib/steps/representation/identifier/index.ts
@@ -1,19 +1,21 @@
 import { ScenarioStep } from '../../index'
-import { HydraResource } from 'alcaeus/types/Resources'
+import { HydraResource } from 'alcaeus/Resources'
 import { E2eContext } from '../../../../types'
 import { checkChain, Result } from 'hydra-validator-core'
+import { NamedNode } from 'rdf-js'
+import { namedNode } from '@rdfjs/data-model'
 
 interface IdentifierStepInit {
   value: string
 }
 
 export class IdentifierStep extends ScenarioStep<HydraResource> {
-  private readonly __identifier: string
+  private readonly __identifier: NamedNode
 
   public constructor(init: IdentifierStepInit) {
     super([])
 
-    this.__identifier = init.value
+    this.__identifier = namedNode(init.value)
   }
 
   protected appliesToInternal(): boolean {
@@ -27,10 +29,10 @@ export class IdentifierStep extends ScenarioStep<HydraResource> {
 
       if (typeof obj !== 'object') {
         result = Result.Failure(`Expected <${__identifier}> resource but the value is a ${typeof obj}`)
-      } else if (__identifier === obj.id) {
+      } else if (__identifier.equals(obj.id)) {
         result = Result.Success(`Found expected resource identifier ${__identifier}`)
       } else {
-        result = Result.Failure(`Expect resource <${__identifier}> but got <${obj.id}> instead.`)
+        result = Result.Failure(`Expect resource <${__identifier}> but got <${obj.id.value}> instead.`)
       }
 
       return {